+++ core/src/main/scala/com/lomekwi/cave/util/Ranges.scala
@@ -4,6 +4,6 @@ import com.google.common.collect.Range
 
 object Ranges {
   def shift(r: Range[java.lang.Long], delta: Long): Range[java.lang.Long] = {
-    Range.closedOpen(java.lang.Long.valueOf(r.lowerEndpoint() + delta), java.lang.Long.valueOf(r.upperEndpoint() + delta))
+    Range.closedOpen(r.lowerEndpoint() + delta, r.upperEndpoint() + delta)
   }
 }
