+++ core/src/test/scala/com/lomekwi/cave/ui/editpanel/tlarea/TlGroupDragSimTest.scala
@@ -73,7 +73,7 @@ class TlGroupDragSimTest extends GdxTestBase {
 
   /** 在模型上放置一个片段，并按 act() 的重建逻辑摆好 Actor。 */
   private def place(track: Track, s: Segment, start: Long, end: Long): SegActor = {
-    timeline.tryAdd(track, s, Range.closedOpen(java.lang.Long.valueOf(start), java.lang.Long.valueOf(end)))
+    timeline.tryAdd(track, s, Range.closedOpen(start, end))
     val actor = s.getActor()
     actor.tl = tl
     actor.setPosition(absX(start), trackTopY(track.index))
@@ -110,13 +110,13 @@ class TlGroupDragSimTest extends GdxTestBase {
 
     // 事件驱动一次：应恰好移动 100_000µs
     actor.dragTo(mouseLocalX - actor.getX, mouseLocalY - actor.getY)
-    assertEquals(Range.closedOpen(java.lang.Long.valueOf(100_000L), java.lang.Long.valueOf(1000_000L + 100_000L)), s.getRange())
+    assertEquals(Range.closedOpen(100_000L, 1000_000L + 100_000L), s.getRange())
 
     // 鼠标不动，多帧重建（纯投影）：模型与 origin 必须稳定
     var i = 0
     while (i < 5) {
       rebuildFromModel(actor)
-      assertEquals(Range.closedOpen(java.lang.Long.valueOf(100_000L), java.lang.Long.valueOf(1000_000L + 100_000L)), s.getRange())
+      assertEquals(Range.closedOpen(100_000L, 1000_000L + 100_000L), s.getRange())
       assertEquals(5_000_000L + 100_000L, s.getOrigin())
       i += 1
     }
@@ -124,7 +124,7 @@ class TlGroupDragSimTest extends GdxTestBase {
     actor.finishDrag()
 
     project.undoManager.undo()
-    assertEquals(Range.closedOpen(java.lang.Long.valueOf(0L), java.lang.Long.valueOf(1000_000L)), s.getRange())
+    assertEquals(Range.closedOpen(0L, 1000_000L), s.getRange())
     assertEquals(5_000_000L, s.getOrigin())
   }
 
@@ -178,14 +178,14 @@ class TlGroupDragSimTest extends GdxTestBase {
 
     val newFrontLocalX = 50f // 起点右移 50px=50ms
     actor.dragTo(newFrontLocalX, view.trackHeight / 2)
-    assertEquals(Range.closedOpen(java.lang.Long.valueOf(50_000L), java.lang.Long.valueOf(1000_000L)), s.getRange())
+    assertEquals(Range.closedOpen(50_000L, 1000_000L), s.getRange())
     assertEquals(0, s.getOrigin())
 
     // 鼠标不动（停在裁切后的新起点 absX(50000)），多帧重建：起点与 origin 都必须稳定
     var i = 0
     while (i < 5) {
       rebuildFromModel(actor)
-      assertEquals(Range.closedOpen(java.lang.Long.valueOf(50_000L), java.lang.Long.valueOf(1000_000L)), s.getRange())
+      assertEquals(Range.closedOpen(50_000L, 1000_000L), s.getRange())
       assertEquals(0, s.getOrigin())
       i += 1
     }
@@ -203,14 +203,14 @@ class TlGroupDragSimTest extends GdxTestBase {
 
     val newWidth = actor.getWidth + 60f // 终点右移 60px=60ms
     actor.dragTo(newWidth, view.trackHeight / 2)
-    assertEquals(Range.closedOpen(java.lang.Long.valueOf(0L), java.lang.Long.valueOf(1000_000L + 60_000L)), s.getRange())
+    assertEquals(Range.closedOpen(0L, 1000_000L + 60_000L), s.getRange())
     assertEquals(5_000_000L, s.getOrigin())
 
     // 鼠标不动，多帧重建：终点与 origin 都必须稳定
     var i = 0
     while (i < 5) {
       rebuildFromModel(actor)
-      assertEquals(Range.closedOpen(java.lang.Long.valueOf(0L), java.lang.Long.valueOf(1000_000L + 60_000L)), s.getRange())
+      assertEquals(Range.closedOpen(0L, 1000_000L + 60_000L), s.getRange())
       assertEquals(5_000_000L, s.getOrigin())
       i += 1
     }
